requirements.txt
alembic==1.16.4
blinker==1.9.0
click==8.1.8
Deprecated==1.2.18
flask==3.1.1
flask-cors==6.0.1
Flask-DotEnv==0.1.2
Flask-JWT-Extended==4.7.1
flask-limiter==3.11.0
flask-mail==0.10.0
Flask-Migrate==4.1.0
flask-sqlalchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
limits==4.2
mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
rich==13.9.4
sqlalchemy==2.0.42
tomli==2.2.1
typing-extensions==4.14.1
werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
psycopg2-binary==2.9.9
matplotlib
pillow
reportlab
python-escpos==2.2.0
